infrared revisions per #14
a) rewrote the English in the Description.docx to be more explanatory
of the overall methods
b) added some comments to both the C and Python source to explain the
interrupt handling logic
c) revised the filenames of corresponding (identical) in both folders
to be the same, in all cases simply "infrared.*" -- infrared.jpg,
infrared.docx, infrared.c, etc.
- fixed a few typos / formatting probs in earlier commits
