+++ K47 User Manual/13_infraredReceiver/Connection method/Pin position corresponding table.txt
@@ -1,12 +1,13 @@
-IR receiver pin position:
-	"S" ====> 10
-	"+" ====> +5V
-	"-" ====> GND
+Infrared Receiver pin position:
+	"S" ====> Raspberry Pi pin 10
+	"+" ====> Raspberry Pi +5V
+	"-" ====> Raspberry Pi GND
 	
-IR emitter pin position:
-	"S" ====> 11
-	"+" ====> +5V
-	"-" ====> GND
-Led pin position:
-	"S" ====> 16
-	"-" ===> GND
+Infrared Emitter pin position:
+	"S" ====> Raspberry Pi 11
+	"+" ====> Raspberry Pi +5V
+	"-" ====> Raspberry Pi GND
+	
+LED pin position:
+	"S" ====> Raspberry Pi 16
+	"-" ===> Raspberry Pi GND
