+++ K47 User Manual/13_infraredReceiver/Connection method/infrared.jpg
