+++ K47 User Manual/14_infraredEmitter/Connection method/infrared.cdr
